renamed junk to trash
synonyms are cool
